+++ algorithms/__init__.py
@@ -1,0 +1,11 @@
+from algorithms.bogoSort import bogoSort
+from algorithms.bubbleSort import bubbleSort
+from algorithms.cocktailSort import cocktailSort
+from algorithms.countingSort import countingSort
+from algorithms.heapSort import heapSort
+from algorithms.insertionSort import insertionSort
+from algorithms.mergeSort import mergeSort
+from algorithms.quickSort import quickSort
+from algorithms.selectionSort import selectionSort
+
+__all__ = ['bogoSort', 'bubbleSort', 'cocktailSort', 'countingSort', 'heapSort', 'insertionSort', 'mergeSort', 'quickSort', 'selectionSort']
